requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.0.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.2
colorama==0.4.6
cryptography==44.0.2
dj-database-url==2.2.0
Django==3.2.25
django-allauth==0.63.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.5.0
django-environ==0.12.0
django-redis==5.4.0
django-redis-cache==3.0.1
django-secure==1.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-yasg==1.21.8
eventlet==0.39.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==3.5.3
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
